Relocate some android publishing files
